move sick note extension stuff into separate package
