+++ src/main/java/org/synyx/urlaubsverwaltung/sicknote/sicknote/extend/SickNoteExtensionEntity.java
@@ -1,4 +1,4 @@
-package org.synyx.urlaubsverwaltung.sicknote.sicknote;
+package org.synyx.urlaubsverwaltung.sicknote.sicknote.extend;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
